web/requirements-web.txt
# RagMind Web — Python dependencies
fastapi>=0.111.0
uvicorn>=0.30.0
python-multipart>=0.0.9   # file uploads

# RAG core
chromadb>=0.5.0
pypdf>=4.0.0
python-docx>=1.1.0
ebooklib>=0.18
beautifulsoup4>=4.12
